requirements.txt
aniso8601==10.0.1
blinker==1.9.0
click==8.2.1
et_xmlfile==2.0.0
Flask==3.1.1
flask-cors==6.0.1
Flask-RESTful==0.3.10
flasgger==0.9.7.1
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.1
MarkupSafe==3.0.2
numpy==2.2.6
openpyxl==3.1.5
pandas==2.3.0
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.3
scikit-learn==1.7.0
scipy==1.15.3
six==1.17.0
threadpoolctl==3.6.0
tzdata==2025.2
Werkzeug==3.1.3
Flask-JWT-Extended==4.6.0
Flask-Bcrypt==1.0.1
python-dotenv==1.0.0
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23
Flask-SQLAlchemy==3.1.1
alembic==1.13.1
flasgger==0.9.7.1
chromadb